requirements.txt
Flask
mysql-connector-python
requests
matplotlib
seaborn
sqlalchemy